// apps/main/src/models/v2/dataplane/byModeQueryParamParameterItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */

export type ByModeQueryParamParameterItem =
  (typeof ByModeQueryParamParameterItem)[keyof typeof ByModeQueryParamParameterItem];

export const ByModeQueryParamParameterItem = {
  plan_only: 'plan_only',
  deploy: 'deploy',
  rollback: 'rollback',
  rollback_plan: 'rollback_plan',
  destroy: 'destroy',
} as const;